Fix scripts

+++ scripts/add-additional-disk.sh
@@ -7,9 +7,20 @@ echo ">>> Add additional disk"
 DISK="/dev/vdb"
 MOUNTPOINT="/mnt/disk"
 
-echo -e "n\np\n1\n\n\nw" | fdisk ${DISK}
-
-mkfs.${FILESYSTEM} ${DISK}1
 mkdir -p ${MOUNTPOINT}
 
-mount -t ${FILESYSTEM} ${DISK}1 ${MOUNTPOINT}
+if [ x"$FILESYSTEM" = x"zfs" ]; then
+  ZFS_POOL_NAME=zfs-vdb
+
+  sudo apt-get install -y zfsutils-linux
+  sudo zpool create ${ZFS_POOL_NAME} ${DISK}
+
+  sudo zfs set mountpoint=${MOUNTPOINT} ${ZFS_POOL_NAME}
+else
+  echo -e "n\np\n1\n\n\nw" | fdisk ${DISK}
+
+  mkfs.${FILESYSTEM} ${DISK}1
+  mkdir -p ${MOUNTPOINT}
+
+  mount -t ${FILESYSTEM} ${DISK}1 ${MOUNTPOINT}
+fi
